+++ backend-api/src/main/java/org/example/backendapi/service/AulaService.java
@@ -37,12 +37,12 @@ public class AulaService {
     private final AulaMapper aulaMapper;
     private final UsuarioMapper usuarioMapper;
 
-    public List<AulaResponseDTO> obtenerAulasPorProfesor(Integer profesorId) {
+    public List<AulaResponseDTO> obtenerAulasPorProfesor(Long profesorId) {
         List<Aula> aulas = aulaDAO.findAulasByProfesorId(profesorId);
         return aulaMapper.toResponseDTOList(aulas);
     }
 
-    public List<UsuarioResponseDTO> obtenerAlumnosPorAula(Integer aulaId) {
+    public List<UsuarioResponseDTO> obtenerAlumnosPorAula(Long aulaId) {
         Aula aula = aulaDAO.findById(aulaId)
                 .orElseThrow(() -> new ResourceNotFoundException("Aula no encontrada"));
 
@@ -68,7 +68,7 @@ public class AulaService {
     }
 
     @Transactional
-    public List<CredencialesResponseDTO> generarAlumnosParaAula(Integer aulaId, Integer cantidad) {
+    public List<CredencialesResponseDTO> generarAlumnosParaAula(Long aulaId, Integer cantidad) {
         Aula aula = aulaDAO.findById(aulaId)
                 .orElseThrow(() -> new ResourceNotFoundException("Aula no encontrada"));
 
@@ -86,7 +86,7 @@ public class AulaService {
     }
 
     @Transactional
-    public List<CredencialesResponseDTO> importarAlumnosCSV(Integer aulaId, MultipartFile file) {
+    public List<CredencialesResponseDTO> importarAlumnosCSV(Long aulaId, MultipartFile file) {
         Aula aula = aulaDAO.findById(aulaId)
                 .orElseThrow(() -> new ResourceNotFoundException("Aula no encontrada"));
